Moved static method GenerateGroupsWithProgress from CharacterGroupGenerator (which is headed for GlyssenEngine) to GenerateGroupsProgressDialog.
